@@ -10,34 +10,36 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "@types/lodash": "^4.14.170",
-    "@types/prismjs": "^1.16.5",
-    "@types/react-dom": "^17.0.8",
+    "@types/lodash": "^4.14.172",
+    "@types/prismjs": "^1.16.6",
+    "@types/react-dom": "^17.0.9",
     "bulma": "^0.9.3",
-    "css-loader": "^5.2.6",
+    "css-loader": "^6.2.0",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.2",
     "raw-loader": "^4.0.2",
     "sass": "^1.32.13",
     "sass-loader": "^12.1.0",
-    "style-loader": "^3.0.0",
-    "ts-loader": "^9.2.3",
-    "typescript": "^4.3.4",
-    "webpack": "^5.40.0",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2"
+    "style-loader": "^3.2.1",
+    "ts-loader": "^9.2.5",
+    "typescript": "^4.4.2",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.52.1",
+    "webpack-cli": "^4.8.0",
+    "webpack-dev-server": "^4.2.0"
   },
   "dependencies": {
     "@egjs/hammerjs": "^2.0.17",
-    "@fortawesome/fontawesome-free": "^5.15.3",
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "axios": "^0.21.1",
+    "@fortawesome/fontawesome-free": "^5.15.4",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-brands-svg-icons": "^5.15.4",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "axios": "^0.21.4",
+    "component-emitter": "^1.3.0",
     "dayjs": "^1.10.6",
     "keycharm": "^0.4.0",
     "lodash": "^4.17.21",
-    "prismjs": "^1.24.0",
+    "prismjs": "^1.24.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "router5": "^8.0.1",
@@ -45,7 +47,7 @@
     "timsort": "^0.3.0",
     "uuid": "^8.3.2",
     "vis-data": "^7.1.2",
-    "vis-network": "^9.0.4",
+    "vis-network": "^9.1.0",
     "vis-util": "^5.0.2"
   }
 }
