@@ -1,9 +1,8 @@
 import { HtmlBlueprint, iterable, mix, observable, patch } from "@chorda/core"
 import { RowLayout } from "chorda-bulma"
-import { Coerced, DataScope, withScope } from "../../utils"
+import { Coerced, DataScope, withItem, withIterableItems, withScope } from "../../utils"
 import { COUNTRIES, Country } from "../../data"
 import { Option, Select } from "./common/select"
-import { withItem, withIterableItems } from "../list/common/utils"
 
 
 export default () : HtmlBlueprint => {
